pyproject.toml
[tool.poetry]
name = "clab-sonic"
version = "0.1.0"
description = "A SONiC Container Lab Playground."
authors = ["Hugo Tinoco <hugotinoco@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.3"
pyyaml = "^6.0.2"
loguru = "^0.7.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
